requirements.txt
flask
gunicorn
flask-sqlalchemy
python-dotenv  # Ensures Flask loads .env correctly
importlib-metadata
